requirements.txt
dataclasses==0.8
Django==2.2.13
future==0.18.2
imageio==2.9.0
numpy==1.19.4
Pillow==8.0.1
pytz==2020.4
sqlparse==0.4.1
torch==1.7.0
torchvision==0.8.1
typing-extensions==3.7.4.3
